feat: add parser fuzz scaffold and CLI/test extraction start
