+++ Cargo.toml
@@ -13,6 +13,7 @@ members = [
     "apps/nec-gui",
     "apps/nec-tui",
 ]
+exclude = ["fuzz"]
 resolver = "2"
 
 [workspace.package]
