+++ fuzz/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "fnec-fuzz"
+version = "0.0.0"
+publish = false
+edition = "2021"
+
+[package.metadata]
+cargo-fuzz = true
+
+[dependencies]
+libfuzzer-sys = "0.4"
+nec_parser = { path = "../crates/nec_parser" }
+
+[[bin]]
+name = "nec_parser_parse"
+path = "fuzz_targets/nec_parser_parse.rs"
+test = false
+doc = false
+bench = false
