+++ fuzz/corpus/nec_parser_parse/basic-dipole.nec
@@ -1,0 +1,5 @@
+GW 1 51 0 0 -5.282 0 0 5.282 0.001
+GE
+EX 0 1 26 0 1.0 0.0
+FR 0 1 0 0 14.2 0.0
+EN
